renamed files to proper format
