+++ src/containers/shared/components/Transaction/AMM/AMMBid/Simple.tsx
