+++ src/containers/shared/components/Transaction/AMM/AMMBid/index.ts
@@ -1,4 +1,4 @@
-import { Simple } from './simple'
+import { Simple } from './Simple'
 import { TransactionMapping } from '../../types'
 import { parser } from './parser'
