+++ src/containers/shared/components/Transaction/AMM/AMMBid/parser.ts
