+++ src/containers/shared/components/Transaction/AMM/AMMCreate/index.ts
@@ -1,9 +1,8 @@
-import { Simple } from './simple'
+import { Simple } from './Simple'
 import { TransactionMapping } from '../../types'
-
 import { parser } from './parser'
 
-export const AMMVote: TransactionMapping = {
+export const AMMCreate: TransactionMapping = {
   Simple,
   parser,
 }
