+++ src/containers/shared/components/Transaction/AMM/AMMDeposit/test/AMMDeposit.test.tsx
@@ -1,4 +1,4 @@
-import { Simple } from '../simple'
+import { Simple } from '../Simple'
 import { createSimpleWrapperFactory, expectSimpleRowText } from '../../../test'
 import depositBothAssets from './mock_data/deposit_both.json'
 import depositUSD from './mock_data/deposit_usd.json'
