+++ src/containers/shared/components/Transaction/AMM/AMMVote/Simple.tsx
