+++ src/containers/shared/components/Transaction/AMM/AMMVote/index.ts
@@ -1,8 +1,9 @@
-import { Simple } from './simple'
+import { Simple } from './Simple'
 import { TransactionMapping } from '../../types'
+
 import { parser } from './parser'
 
-export const AMMDeposit: TransactionMapping = {
+export const AMMVote: TransactionMapping = {
   Simple,
   parser,
 }
