+++ src/containers/shared/components/Transaction/AMM/AMMVote/parser.ts
